confirmation message hooks

+++ components/data-table.tsx
@@ -24,22 +24,26 @@ import {
 import { Button } from "./ui/button";
 import { Input } from "@/components/ui/input";
 import { Trash } from "lucide-react";
-
+import useConfirm from "@/hooks/use-confirm";
 interface DataTableProps<TData, TValue> {
   columns: ColumnDef<TData, TValue>[];
   data: TData[];
   filterKey: string;
-  onDelete: (rows:Row<TData>[]) => void, 
-  disabled?: boolean
+  onDelete: (rows: Row<TData>[]) => void;
+  disabled?: boolean;
 }
 
 export function DataTable<TData, TValue>({
   columns,
   data,
   filterKey,
-  onDelete, 
-  disabled, 
+  onDelete,
+  disabled,
 }: DataTableProps<TData, TValue>) {
+  const [ConfirmDialog, confirm] = useConfirm(
+    "Are you sure?",
+    "This will delete many accounts"
+  );
   const [sorting, setSorting] = React.useState<SortingState>([]);
   const [columnFilters, setColumnFilters] = React.useState<ColumnFiltersState>(
     []
@@ -64,6 +68,7 @@ export function DataTable<TData, TValue>({
 
   return (
     <div>
+      <ConfirmDialog />
       <div className="flex items-center py-4">
         <Input
           placeholder={`Filter ${filterKey}...`}
@@ -75,19 +80,23 @@ export function DataTable<TData, TValue>({
           }
           className="max-w-sm"
         />
-        {table.getFilteredSelectedRowModel().rows.length > 0  && (
-            <Button size={"sm" } variant={"outline"}
-                className="ml-auto font-normal text-xs"
-                onClick={() => {
-                  onDelete(table.getFilteredRowModel().rows)
-                  table.resetRowSelection()
-                }}
-            >
-                <Trash
-                 className="size-4 mr-2"
-                />
-                Delete ({table.getFilteredSelectedRowModel().rows.length})
-            </Button>
+        {table.getFilteredSelectedRowModel().rows.length > 0 && (
+          <Button
+            size={"sm"}
+            variant={"outline"}
+            className="ml-auto font-normal text-xs"
+            onClick={async () => {
+              const ok = await confirm();
+
+              if (ok) {
+                onDelete(table.getFilteredRowModel().rows);
+                table.resetRowSelection();
+              }
+            }}
+          >
+            <Trash className="size-4 mr-2" />
+            Delete ({table.getFilteredSelectedRowModel().rows.length})
+          </Button>
         )}
       </div>
       <div className="rounded-md border">
